rename schema to model
