+++ app/days/controllers/find_by_month.js
@@ -1,6 +1,6 @@
 var express = require('express')
 var router = express.Router()
-var Meal = require('../schema.js')
+var Meal = require('../model.js')
 
 router.get('/:month', function (req,res) {
   Meal.find({}, function(err, meals) {
