+++ app/days/controllers/find_by_week.js
@@ -1,6 +1,6 @@
 var express = require('express')
 var router = express.Router()
-var Meal = require('../schema.js')
+var Meal = require('../model.js')
 
 router.get('/:week', function (req,res) {
   Meal.find({}, function(err, meals) {
