+++ app/meals/controllers/find_by_id.js
@@ -1,6 +1,6 @@
 var express = require('express')
 var router = express.Router()
-var Meal = require('../schema.js')
+var Meal = require('../model.js')
 
 router.get('/:id', function (req,res) {
   var id = req.param("id")
