+++ app/users/controllers/create.js
@@ -1,6 +1,6 @@
 var express = require('express')
 var router = express.Router()
-var User = require('../schema.js')
+var User = require('../model.js')
 
 router.post('/', function (req,res) {
   User.create(req.body, function (err, newUser) {
