+++ app/users/controllers/find_by_id.js
@@ -1,6 +1,6 @@
 var express = require('express')
 var router = express.Router()
-var User = require('../schema.js')
+var User = require('../model.js')
 
 router.get('/:id', function (req,res) {
   var id = req.param("id")
